logging: cast log file max_size to size_t explicitly
airy_log_init() assigns the uint64_t file max_size into the size_t
log_config_t max_file_size field. On windows-x86-32 size_t is 32-bit,
so MSVC raises C4244 (possible loss of data) which /WX promotes to
error C2220 (probe-3 evidence, run 34200768597).

Add an explicit narrowing cast: log-file size caps above 4 GiB are not
representable on that ABI and truncation is the intended behaviour.
On x64 and LP64 platforms the cast is a no-op.

Fixes: 8674e3d48cc4 ("chore: fix copyright year 2026→2025-2026 and NOTICE URL")

+++ utils/logging/svc_logger.h
@@ -169,7 +169,7 @@ static inline int airy_log_init(const airy_logger_config_t *config)
             if (config->targets[i].type == AIRY_LOG_TARGET_FILE) {
                 log_cfg.outputs |= (1 << LOG_OUTPUT_FILE);
                 log_cfg.file_path = config->targets[i].config.file.path;
-                log_cfg.max_file_size = config->targets[i].config.file.max_size;
+                log_cfg.max_file_size = (size_t)config->targets[i].config.file.max_size;
                 log_cfg.max_backup_count = config->targets[i].config.file.max_files;
             }
         }
